fix(publish dialog): preselection of Status and Labels field for republish attempt (#69)
Fixed working with defaultvalue in the React Hook Form. Now the uploaded values update the form.
Fixed the list of available items in the dropdown list, now all available options are shown there, even if one of them is already selected.
The filtering and item selection cases are separated. It should increase reliability.
Clearing the status field is prohibited.
Added a feature to disable dependent fields from the workflow or package. Until the workflow or package is selected, the public fields will not be available for editing.
Added disabling field editing when loading defaults from the backend.
Added the ability to disable labels
Disabled status chips display has been fixed